Leverages the ops testing framework in the `provides` part of the relation
