+++ tests/charms/tls_certificates_interface/v1/test_tls_certificates_v1.py
@@ -2,14 +2,10 @@
 # Copyright 2021 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import json
-import unittest
 import uuid
-from unittest.mock import Mock, PropertyMock, call, patch
 
 import pytest
 from charms.tls_certificates_interface.v1.tls_certificates import (
-    TLSCertificatesProvidesV1,
     generate_ca,
     generate_certificate,
     generate_csr,
@@ -33,73 +29,6 @@ from tests.charms.tls_certificates_interface.v1.certificates import (
 from tests.charms.tls_certificates_interface.v1.certificates import (
     generate_private_key as generate_private_key_helper,
 )
-
-PROVIDER_UNIT_NAME = "whatever provider unit name"
-PROVIDER_APP_NAME = "whatever provider app name"
-REQUIRER_UNIT_NAME = "whatever requirer unit name"
-CHARM_LIB_PATH = "charms.tls_certificates_interface.v1.tls_certificates"
-
-
-class RelationMock:
-    def __init__(
-        self,
-        provider_app,
-        requirer_unit,
-        relation_id: int = 123,
-        provider_app_data: dict = None,
-        requirer_unit_data: dict = None,
-    ):
-        if provider_app_data:
-            self.provider_app_data = provider_app_data
-        else:
-            self.provider_app_data = dict()
-
-        if requirer_unit_data:
-            self.requirer_unit_data = requirer_unit_data
-        else:
-            self.requirer_unit_data = dict()
-        self.provider_app = provider_app
-        self.requirer_unit = requirer_unit
-        self._relation_id = relation_id
-
-    @property
-    def data(self):
-        return {
-            self.provider_app: self.provider_app_data,
-            self.requirer_unit: self.requirer_unit_data,
-        }
-
-    @property
-    def app(self):
-        return self.provider_app
-
-    @property
-    def units(self):
-        return [self.requirer_unit]
-
-    @property
-    def id(self):
-        return self._relation_id
-
-
-def _load_relation_data(raw_relation_data: dict) -> dict:
-    """Loads relation data from the relation data bag.
-
-    Json loads all data.
-
-    Args:
-        raw_relation_data: Relation data from the databag
-
-    Returns:
-        dict: Relation data in dict format.
-    """
-    certificate_data = dict()
-    for key in raw_relation_data:
-        try:
-            certificate_data[key] = json.loads(raw_relation_data[key])
-        except json.decoder.JSONDecodeError:
-            certificate_data[key] = raw_relation_data[key]
-    return certificate_data
 
 
 def validate_induced_data_from_pfx_is_equal_to_initial_data(
@@ -139,404 +68,6 @@ def validate_induced_data_from_pfx_is_equal_to_initial_data(
     assert initial_public_key == induced_public_key
     assert induced_certificate == initial_certificate
     assert initial_private_key == induced_private_key
-
-
-class UnitMock:
-    def __init__(self, name):
-        self.name = name
-
-    @staticmethod
-    def is_leader():
-        return True
-
-
-class AppMock:
-    def __init__(self, name):
-        self.name = name
-
-
-class TestTLSCertificatesProvidesV1(unittest.TestCase):
-    def setUp(self):
-        class MockRelationEvents:
-            def relation_changed(self):
-                pass
-
-        self.relationship_name = "certificates"
-        charm = Mock()
-        charm.on = {"certificates": MockRelationEvents()}
-        self.tls_relation_provides = TLSCertificatesProvidesV1(
-            charm=charm, relationship_name=self.relationship_name
-        )
-        self.charm = charm
-        self.provider_app = AppMock(name=PROVIDER_APP_NAME)
-        self.requirer_unit = UnitMock(name=REQUIRER_UNIT_NAME)
-        self.charm.framework.model.app = self.provider_app
-
-    @patch(
-        f"{CHARM_LIB_PATH}.CertificatesProviderCharmEvents.certificate_creation_request",
-        new_callable=PropertyMock,
-    )
-    def test_given_csr_in_relation_data_when_relation_changed_then_certificate_creation_request_is_emitted(  # noqa: E501
-        self, patch_certificate_creation_request
-    ):
-        event = Mock()
-        csr = "whatever csr"
-        relation_id = 456
-        self.charm.framework.model.get_relation.return_value = event.relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            requirer_unit_data={
-                "certificate_signing_requests": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": csr,
-                        }
-                    ]
-                )
-            },
-            relation_id=relation_id,
-        )
-        event.unit = self.requirer_unit
-
-        self.tls_relation_provides._on_relation_changed(event)
-
-        patch_certificate_creation_request.assert_has_calls(
-            [call().emit(certificate_signing_request=csr, relation_id=relation_id)]
-        )
-
-    @patch(
-        f"{CHARM_LIB_PATH}.CertificatesProviderCharmEvents.certificate_creation_request",
-        new_callable=PropertyMock,
-    )
-    def test_given_no_csr_in_certificate_signing_request_when_relation_changed_then_certificate_creation_request_is_not_emitted(  # noqa: E501
-        self, patch_certificate_creation_request
-    ):
-        event = Mock()
-        self.charm.framework.model.get_relation.return_value = event.relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            requirer_unit_data={
-                "certificate_signing_requests": json.dumps(
-                    [
-                        {
-                            "invalid key": "invalid value",
-                        }
-                    ]
-                )
-            },
-        )
-        event.unit = self.requirer_unit
-
-        self.tls_relation_provides._on_relation_changed(event)
-
-        patch_certificate_creation_request.assert_not_called()
-
-    @patch(
-        f"{CHARM_LIB_PATH}.CertificatesProviderCharmEvents.certificate_creation_request",
-        new_callable=PropertyMock,
-    )
-    def test_given_certificate_for_csr_already_in_relation_data_when_on_relation_changed_then_certificate_creation_request_is_not_emitted(  # noqa: E501
-        self, patch_certificate_creation_request
-    ):
-        csr = "whatever csr"
-        event = Mock()
-        self.charm.framework.model.get_relation.return_value = event.relation = RelationMock(
-            requirer_unit=self.requirer_unit,
-            provider_app=self.provider_app,
-            requirer_unit_data={
-                "certificate_signing_requests": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": csr,
-                        }
-                    ]
-                )
-            },
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": csr,
-                            "certificate": "whatever cert",
-                            "ca": "whatever ca",
-                            "chain": ["whatever cert 1", "whatever cert 2"],
-                        }
-                    ]
-                )
-            },
-        )
-        event.unit = self.requirer_unit
-
-        self.tls_relation_provides._on_relation_changed(event)
-
-        patch_certificate_creation_request.assert_not_called()
-
-    @patch(f"{CHARM_LIB_PATH}.TLSCertificatesProvidesV1.remove_certificate")
-    @patch(
-        f"{CHARM_LIB_PATH}.CertificatesProviderCharmEvents.certificate_revocation_request",
-    )
-    def test_given_csr_in_provider_relation_data_but_not_in_requirer_when_on_relation_changed_then_certificate_revocation_request_is_emitted(  # noqa: E501
-        self, patch_certificate_revocation_request, _
-    ):
-        event = Mock()
-        certificate = "whatever cert"
-        csr = "whatever csr"
-        ca = "whatever ca"
-        chain = ["whatever cert 1", "whatever cert 2"]
-        self.charm.framework.model.get_relation.return_value = event.relation = RelationMock(
-            requirer_unit=self.requirer_unit,
-            provider_app=self.provider_app,
-            requirer_unit_data={"certificate_signing_requests": json.dumps([])},
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": csr,
-                            "certificate": certificate,
-                            "ca": ca,
-                            "chain": chain,
-                        }
-                    ]
-                )
-            },
-        )
-        event.unit = self.requirer_unit
-
-        self.tls_relation_provides._on_relation_changed(event)
-
-        patch_certificate_revocation_request.emit.assert_called_with(
-            certificate=certificate,
-            certificate_signing_request=csr,
-            ca=ca,
-            chain=chain,
-        )
-
-    @patch(f"{CHARM_LIB_PATH}.TLSCertificatesProvidesV1.remove_certificate")
-    @patch(
-        f"{CHARM_LIB_PATH}.CertificatesProviderCharmEvents.certificate_revocation_request",
-        new_callable=PropertyMock,
-    )
-    def test_given_csr_in_provider_relation_data_but_not_in_requirer_when_on_relation_changed_then_remove_certificate_is_called(  # noqa: E501
-        self,
-        _,
-        patch_remove_certificate,
-    ):
-        certificate = "whatever cert"
-        event = Mock()
-        self.charm.framework.model.get_relation.return_value = event.relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": "whatever csr",
-                            "certificate": certificate,
-                            "ca": "whatever ca",
-                            "chain": ["whatever cert 1", "whatever cert 2"],
-                        }
-                    ]
-                )
-            },
-            requirer_unit_data={"certificate_signing_requests": json.dumps([])},
-        )
-        event.unit = self.requirer_unit
-
-        self.tls_relation_provides._on_relation_changed(event)
-
-        patch_remove_certificate.assert_called_with(certificate=certificate)
-
-    def test_given_no_data_in_relation_data_when_set_relation_certificate_then_certificate_is_added_to_relation_data(  # noqa: E501
-        self,
-    ):
-        ca = "whatever ca"
-        certificate = "whatever certificate"
-        certificate_signing_request = "whatever certificate signing request"
-        chain = ["whatever cert 1", "whatever cert 2"]
-        relation_id = 123
-        relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            relation_id=relation_id,
-        )
-        self.charm.framework.model.get_relation.return_value = relation
-
-        self.tls_relation_provides.set_relation_certificate(
-            certificate=certificate,
-            ca=ca,
-            chain=chain,
-            certificate_signing_request=certificate_signing_request,
-            relation_id=relation_id,
-        )
-
-        expected_relation_data = {
-            "certificates": [
-                {
-                    "certificate": certificate,
-                    "certificate_signing_request": certificate_signing_request,
-                    "ca": ca,
-                    "chain": chain,
-                }
-            ]
-        }
-        loaded_relation_data = _load_relation_data(relation.data[self.provider_app])
-        self.assertEqual(expected_relation_data, loaded_relation_data)
-
-    def test_given_some_certificates_in_relation_data_when_set_relation_certificate_then_certificate_is_added_to_relation_data(  # noqa: E501
-        self,
-    ):
-        initial_certificate = "whatever initial cert"
-        initial_certificate_signing_request = "whatever initial csr"
-        initial_ca = "whatever initial ca"
-        initial_chain = ["whatever initial cert 1", "whatever initial cert 2"]
-        new_ca = "whatever new ca"
-        new_certificate = "whatever new certificate"
-        new_certificate_signing_request = "whatever new certificate signing request"
-        new_chain = ["whatever new cert 1", "whatever new cert 2"]
-        relation_id = 123
-        relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": initial_certificate_signing_request,
-                            "certificate": initial_certificate,
-                            "ca": initial_ca,
-                            "chain": initial_chain,
-                        }
-                    ]
-                )
-            },
-        )
-        self.charm.framework.model.get_relation.return_value = relation
-
-        self.tls_relation_provides.set_relation_certificate(
-            certificate=new_certificate,
-            ca=new_ca,
-            chain=new_chain,
-            certificate_signing_request=new_certificate_signing_request,
-            relation_id=relation_id,
-        )
-
-        expected_relation_data = {
-            "certificates": [
-                {
-                    "certificate": initial_certificate,
-                    "certificate_signing_request": initial_certificate_signing_request,
-                    "ca": initial_ca,
-                    "chain": initial_chain,
-                },
-                {
-                    "certificate": new_certificate,
-                    "certificate_signing_request": new_certificate_signing_request,
-                    "ca": new_ca,
-                    "chain": new_chain,
-                },
-            ]
-        }
-        loaded_relation_data = _load_relation_data(relation.data[self.provider_app])
-        self.assertEqual(expected_relation_data, loaded_relation_data)
-
-    def test_given_identical_csr_in_relation_data_when_set_relation_certificate_then_certificate_is_replaced_in_relation_data(  # noqa: E501
-        self,
-    ):
-        initial_certificate = "whatever initial cert"
-        initial_certificate_signing_request = "whatever initial csr"
-        initial_ca = "whatever initial ca"
-        initial_chain = ["whatever initial cert 1", "whatever initial cert 2"]
-        new_certificate = "whatever new certificate"
-        relation_id = 123
-        relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": initial_certificate_signing_request,
-                            "certificate": initial_certificate,
-                            "ca": initial_ca,
-                            "chain": initial_chain,
-                        }
-                    ]
-                )
-            },
-        )
-        self.charm.framework.model.get_relation.return_value = relation
-
-        self.tls_relation_provides.set_relation_certificate(
-            certificate=new_certificate,
-            ca=initial_ca,
-            chain=initial_chain,
-            certificate_signing_request=initial_certificate_signing_request,
-            relation_id=relation_id,
-        )
-
-        expected_relation_data = {
-            "certificates": [
-                {
-                    "certificate": new_certificate,
-                    "certificate_signing_request": initial_certificate_signing_request,
-                    "ca": initial_ca,
-                    "chain": initial_chain,
-                },
-            ]
-        }
-        loaded_relation_data = _load_relation_data(relation.data[self.provider_app])
-        self.assertEqual(expected_relation_data, loaded_relation_data)
-
-    def test_given_certificate_in_relation_data_when_remove_certificate_then_certificate_is_removed_from_relation(  # noqa: E501
-        self,
-    ):
-        certificate = "whatever cert"
-        self.charm.framework.model.get_relation.return_value = relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            provider_app_data={
-                "certificates": json.dumps(
-                    [
-                        {
-                            "certificate_signing_request": "whatever csr",
-                            "certificate": certificate,
-                            "ca": "whatever ca",
-                            "chain": ["whatever cert 1", "whatever cert 2"],
-                        }
-                    ]
-                )
-            },
-        )
-        self.charm.framework.model.relations = {self.relationship_name: [relation]}
-
-        self.tls_relation_provides.remove_certificate(certificate=certificate)
-
-        provider_relation_data = _load_relation_data(relation.data[self.provider_app])
-        self.assertEqual({"certificates": []}, provider_relation_data)
-
-    def test_given_certificate_not_in_relation_data_when_remove_certificate_then_certificate_is_removed_from_relation(  # noqa: E501
-        self,
-    ):
-        user_provided_certificate = "whatever cert"
-        certificates_in_relation_data = [
-            {
-                "certificate_signing_request": "whatever csr",
-                "certificate": "another certificate",
-                "ca": "whatever ca",
-                "chain": ["whatever cert 1", "whatever cert 2"],
-            }
-        ]
-        self.charm.framework.model.get_relation.return_value = relation = RelationMock(
-            provider_app=self.provider_app,
-            requirer_unit=self.requirer_unit,
-            provider_app_data={"certificates": json.dumps(certificates_in_relation_data)},
-        )
-        self.charm.framework.model.relations = {self.relationship_name: [relation]}
-
-        self.tls_relation_provides.remove_certificate(certificate=user_provided_certificate)
-
-        provider_relation_data = _load_relation_data(relation.data[self.provider_app])
-        self.assertEqual({"certificates": certificates_in_relation_data}, provider_relation_data)
 
 
 def test_given_subject_and_private_key_when_generate_csr_then_csr_is_generated_with_provided_subject():  # noqa: E501
